+++ src/backend/impress/settings.py
@@ -212,7 +212,11 @@ class Base(Configuration):
         "application/x-msdownload",
         "application/xml",
     ]
-
+    DOCUMENT_ATTACHMENT_CHECK_UNSAFE_MIME_TYPES_ENABLED = values.BooleanValue(
+        True,
+        environ_name="DOCUMENT_ATTACHMENT_CHECK_UNSAFE_MIME_TYPES_ENABLED",
+        environ_prefix=None,
+    )
     # Document versions
     DOCUMENT_VERSIONS_PAGE_SIZE = 50