+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to
 - 📝(project) add system-requirement doc #1066
 - 🔧(front) configure x-frame-options to DENY in nginx conf #1084
 -  (doc) add documentation to install with compose #855
+- ✨(backend) allow to disable checking unsafe mimetype on attachment upload
 
 ### Changed